refactor: Abstracted invoice object into class
